For #5905 - Combine theme and toolbar settings
