+++ app/src/main/java/org/mozilla/fenix/settings/CustomizationFragment.kt
@@ -15,21 +15,25 @@ import org.mozilla.fenix.components.metrics.Event
 import org.mozilla.fenix.ext.components
 import org.mozilla.fenix.ext.getPreferenceKey
 import org.mozilla.fenix.ext.requireComponents
+import org.mozilla.fenix.ext.settings
 import org.mozilla.fenix.ext.showToolbar
 
-class ThemeFragment : PreferenceFragmentCompat() {
+/**
+ * Lets the user customize the UI.
+ */
+class CustomizationFragment : PreferenceFragmentCompat() {
     private lateinit var radioLightTheme: RadioButtonPreference
     private lateinit var radioDarkTheme: RadioButtonPreference
     private lateinit var radioAutoBatteryTheme: RadioButtonPreference
     private lateinit var radioFollowDeviceTheme: RadioButtonPreference
 
     override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
-        setPreferencesFromResource(R.xml.theme_preferences, rootKey)
+        setPreferencesFromResource(R.xml.customization_preferences, rootKey)
     }
 
     override fun onResume() {
         super.onResume()
-        showToolbar(getString(R.string.preferences_theme))
+        showToolbar(getString(R.string.preferences_customize))
         setupPreferences()
     }
 
@@ -39,6 +43,7 @@ class ThemeFragment : PreferenceFragmentCompat() {
         bindLightTheme()
         bindAutoBatteryTheme()
         setupRadioGroups()
+        setupToolbarCategory()
     }
 
     private fun setupRadioGroups() {
@@ -110,5 +115,29 @@ class ThemeFragment : PreferenceFragmentCompat() {
             engine.settings.preferredColorScheme = getPreferredColorScheme()
         }
         requireComponents.useCases.sessionUseCases.reload.invoke()
+    }
+
+    private fun setupToolbarCategory() {
+        val keyToolbarTop = getPreferenceKey(R.string.pref_key_toolbar_top)
+        val topPreference = requireNotNull(findPreference<RadioButtonPreference>(keyToolbarTop))
+        topPreference.onClickListener {
+            requireContext().components.analytics.metrics.track(Event.ToolbarPositionChanged(
+                Event.ToolbarPositionChanged.Position.TOP
+            ))
+        }
+
+        val keyToolbarBottom = getPreferenceKey(R.string.pref_key_toolbar_bottom)
+        val bottomPreference = requireNotNull(findPreference<RadioButtonPreference>(keyToolbarBottom))
+        bottomPreference.onClickListener {
+            requireContext().components.analytics.metrics.track(Event.ToolbarPositionChanged(
+                Event.ToolbarPositionChanged.Position.BOTTOM
+            ))
+        }
+
+        topPreference.setCheckedWithoutClickListener(!requireContext().settings().shouldUseBottomToolbar)
+        bottomPreference.setCheckedWithoutClickListener(requireContext().settings().shouldUseBottomToolbar)
+
+        topPreference.addToRadioGroup(bottomPreference)
+        bottomPreference.addToRadioGroup(topPreference)
     }
 }
