+++ app/src/main/res/values/preference_keys.xml
@@ -39,7 +39,7 @@
     <string name="pref_key_sign_in" translatable="false">pref_key_sign_in</string>
     <string name="pref_key_account_auth_error" translatable="false">pref_key_account_auth_error</string>
     <string name="pref_key_private_mode" translatable="false">pref_key_private_mode</string>
-    <string name="pref_key_theme" translatable="false">pref_key_theme</string>
+    <string name="pref_key_customize" translatable="false">pref_key_customize</string>
     <string name="pref_key_toolbar" translatable="false">pref_key_toolbar</string>
     <string name="pref_key_leakcanary" translatable="false">pref_key_leakcanary</string>
     <string name="pref_key_remote_debugging" translatable="false">pref_key_remote_debugging</string>
