--- app/src/main/res/xml/theme_preferences.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- This Source Code Form is subject to the terms of the Mozilla Public
-   - License, v. 2.0. If a copy of the MPL was not distributed with this
-   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
-<androidx.preference.PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <org.mozilla.fenix.settings.RadioButtonPreference
-        android:defaultValue="@bool/underAPI28"
-        android:key="@string/pref_key_light_theme"
-        android:title="@string/preference_light_theme" />
-
-    <org.mozilla.fenix.settings.RadioButtonPreference
-        android:defaultValue="false"
-        android:key="@string/pref_key_dark_theme"
-        android:title="@string/preference_dark_theme" />
-
-    <org.mozilla.fenix.settings.RadioButtonPreference
-        android:defaultValue="false"
-        android:key="@string/pref_key_auto_battery_theme"
-        android:title="@string/preference_auto_battery_theme"
-        app:isPreferenceVisible="@bool/underAPI28" />
-
-    <org.mozilla.fenix.settings.RadioButtonPreference
-        android:defaultValue="@bool/API28"
-        android:key="@string/pref_key_follow_device_theme"
-        android:title="@string/preference_follow_device_theme"
-        app:isPreferenceVisible="@bool/API28" />
-</androidx.preference.PreferenceScreen>
