--- app/src/main/res/xml/toolbar_preferences.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- This Source Code Form is subject to the terms of the Mozilla Public
-   - License, v. 2.0. If a copy of the MPL was not distributed with this
-   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
-    <org.mozilla.fenix.settings.RadioButtonPreference
-        android:key="@string/pref_key_toolbar_top"
-        android:title="@string/preference_top_toolbar" />
-    <org.mozilla.fenix.settings.RadioButtonPreference
-        android:key="@string/pref_key_toolbar_bottom"
-        android:title="@string/preference_bottom_toolbar" />
-</PreferenceScreen>
